quick improvement to the RANSAC steps ordering for the epipolar transformation solver.

+++ src/algorithms/imageProcessing/matching/ORBMatcher.java
@@ -557,7 +557,7 @@ public class ORBMatcher {
         
         RANSACSolver solver = new RANSACSolver();
         
-        int tolerance = 4;
+        double tolerance = 4;
         
         EpipolarTransformationFit fit = solver.calculateEpipolarProjection(
             matchedLeftXY, matchedRightXY, outputLeftXY, outputRightXY,
