+++ src/algorithms/imageProcessing/matching/PartialShapeMatcher.java
@@ -1282,7 +1282,7 @@ public class PartialShapeMatcher {
         
         RANSACSolver solver = new RANSACSolver();
 
-        int tolerance = 4;
+        double tolerance = 4;
         
         EpipolarTransformationFit fit = solver.calculateEpipolarProjection(
             matchedLeftXY, matchedRightXY, outputLeftXY, outputRightXY,
