+++ src/algorithms/imageProcessing/matching/PartialShapeMatcher2.java
@@ -1306,7 +1306,7 @@ public class PartialShapeMatcher2 {
         
         RANSACSolver solver = new RANSACSolver();
 
-        int tolerance = 4;
+        double tolerance = 4;
         
         EpipolarTransformationFit fit = solver.calculateEpipolarProjection(
             matchedLeftXY, matchedRightXY, outputLeftXY, outputRightXY,
